test: extract useAsyncData + useFetch tests
